+++ src/pages/GraduationCalculatorPage/index.tsx
@@ -75,7 +75,7 @@ function GraduationCalculatorPage() {
                 이수구분 등 잘못된 정보를 정정하면 아래의 그래프에 바로 적용돼요.
               </p>
             </div>
-            <GeneralCourse frameId={currentFrameIndex} />
+            <GeneralCourse />
             <CreditChart currentFrameIndex={currentFrameIndex} />
           </div>
         </div>
